+++ Source/Bugsnag.m
@@ -33,6 +33,11 @@
 
 static BugsnagClient *bsg_g_bugsnag_client = NULL;
 
+@interface BugsnagConfiguration ()
+@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
+@property(readwrite, retain, nullable) BugsnagMetadata *config;
+@end
+
 @interface Bugsnag ()
 + (BugsnagClient *)client;
 + (BOOL)bugsnagStarted;