+++ Source/BugsnagConfiguration.h
@@ -32,7 +32,6 @@
 #import "BugsnagMetadata.h"
 #import "BugsnagPlugin.h"
 
-@class BugsnagBreadcrumbs;
 @class BugsnagUser;
 
 /**
@@ -67,18 +66,6 @@ typedef bool (^BugsnagOnSendBlock)(BugsnagEvent *_Nonnull event);
  * @param sessionPayload The session about to be delivered
  */
 typedef void(^BugsnagOnSessionBlock)(NSMutableDictionary *_Nonnull sessionPayload);
-
-/**
- *  A handler for modifying data before sending it to Bugsnag
- *
- *  @param rawEventReports The raw event data written at crash time. This
- *                         includes data added in onError.
- *  @param report          The default report payload
- *
- *  @return the report payload intended to be sent or nil to cancel sending
- */
-typedef NSDictionary *_Nullable (^BugsnagBeforeNotifyHook)(
-    NSArray *_Nonnull rawEventReports, NSDictionary *_Nonnull report);
 
 typedef NS_OPTIONS(NSUInteger, BSGErrorType) {
     BSGErrorTypesNone         NS_SWIFT_NAME(None)         = 0,
@@ -128,21 +115,6 @@ typedef NS_OPTIONS(NSUInteger, BSGErrorType) {
 @property(retain, nullable) BugsnagUser *currentUser;
 
 /**
- *  Additional information about the state of the app or environment at the
- *  time the report was generated
- */
-@property(readwrite, retain, nullable) BugsnagMetadata *metadata;
-/**
- *  Meta-information about the state of Bugsnag
- */
-@property(readwrite, retain, nullable) BugsnagMetadata *config;
-/**
- *  Rolling snapshots of user actions leading up to a crash report
- */
-@property(readonly, strong, nullable)
-BugsnagBreadcrumbs *breadcrumbs;
-
-/**
  *  Optional handler invoked when an error or crash occurs
  */
 @property void (*_Nullable onCrashHandler)
@@ -161,34 +133,9 @@ BugsnagBreadcrumbs *breadcrumbs;
 @property BOOL autoTrackSessions;
 
 /**
- * Whether the app should report out of memory events which terminate the app
- * while the app is in the background. Setting this property has no effect.
- */
-@property BOOL reportBackgroundOOMs
-    __deprecated_msg("This detection option is unreliable and should no longer be used.");
-
-/**
  * The types of breadcrumbs which will be captured. By default, this is all types.
  */
 @property BSGEnabledBreadcrumbType enabledBreadcrumbTypes;
-
-/**
- * Retrieves the endpoint used to notify Bugsnag of errors
- *
- * NOTE: If you want to set this value, you should do so via setEndpointsForNotify:sessions: instead.
- *
- * @see setEndpointsForNotify:sessions:
- */
-@property(readonly, retain, nullable) NSURL *notifyURL;
-
-/**
- * Retrieves the endpoint used to send tracked sessions to Bugsnag
- *
- * NOTE: If you want to set this value, you should do so via setEndpointsForNotify:sessions: instead.
- *
- * @see setEndpointsForNotify:sessions:
- */
-@property(readonly, retain, nullable) NSURL *sessionURL;
 
 @property(retain, nullable) NSString *codeBundleId;
 @property(retain, nullable) NSString *notifierType;
@@ -199,18 +146,6 @@ BugsnagBreadcrumbs *breadcrumbs;
  * messages.
  */
 @property NSUInteger maxBreadcrumbs;
-
-/**
- * Determines whether app sessions should be tracked automatically. By default this value is true.
- * If this value is updated after +[Bugsnag start] is called, only subsequent automatic sessions
- * will be captured.
- */
-@property BOOL shouldAutoCaptureSessions __deprecated_msg("Use autoTrackSessions instead");
-
-/**
- *  YES if uncaught exceptions should be reported automatically
- */
-@property BOOL autoNotify __deprecated_msg("Use autoDetectErrors instead");
 
 /**
  * Whether User information should be persisted to disk between application runs.
@@ -306,16 +241,6 @@ BugsnagBreadcrumbs *breadcrumbs;
  * @param block The block to be removed.
  */
 - (void)removeOnSendBlock:(BugsnagOnSendBlock _Nonnull )block;
-
-/**
- *  Whether reports shoould be sent, based on release stage options
- *
- *  @return YES if reports should be sent based on this configuration
- */
-- (BOOL)shouldSendReports;
-
-- (NSDictionary *_Nonnull)errorApiHeaders;
-- (NSDictionary *_Nonnull)sessionApiHeaders;
 
 - (void)addPlugin:(id<BugsnagPlugin> _Nonnull)plugin;