+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Bugsnag Notifiers on other platforms.
 
 ## Enhancements
 
+* Remove unused APIs from `BugsnagConfiguration` interface
+  [#496](https://github.com/bugsnag/bugsnag-cocoa/pull/496)
+
 * Remove unused APIs from `BugsnagBreadcrumb` interface
   [#502](https://github.com/bugsnag/bugsnag-cocoa/pull/502)